Grid and pagination adjustments
